Refactoring: move option definition and parsing to own files
